+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 
 * \#151 Dropping collections now always uses primary. (Christos Trochalakis)
 
+* \#150 Handle cases where Mongo does not bring back a `query_failure` flag
+  in the reply, but has an error document present.
+
 * mongoid/mongoid#2849 Supply proper limit to initial query if either
   limit or batch_size are provided.